+++ movai_core_shared/logger.py
@@ -46,6 +46,7 @@ from movai_core_shared.envvars import (
     SYSLOG_ENABLED,
     DETACHED_PROCESS_OUTPUT,
 )
+from movai_core_shared.messages.metric_data import MetricQueryResponse
 from movai_core_shared.core.message_client import MessageClient, AsyncMessageClient
 from movai_core_shared.common.utils import is_enterprise, is_manager
 from movai_core_shared.common.time import validate_time
@@ -433,9 +434,8 @@ class LogsQuery:
         toDate=None,
         order_by=None,
         order_dir=None,
-        pagination=False,
         **kwrargs,
-    ):
+    ) -> MetricQueryResponse:
         """Get logs from message-server"""
         server_addr = MASTER_MESSAGE_SERVER
         if is_manager():
@@ -489,7 +489,5 @@ class LogsQuery:
         query_response = await message_client.send_request(
             LOGS_QUERY_HANDLER_MSG_TYPE, query_data, None, True
         )
-        if "response" in query_response:
-            response = query_response["response"]
 
-        return response if pagination else response.get("data", [])
+        return MetricQueryResponse(**(query_response["response"]))