+++ movai_core_shared/messages/metric_data.py
@@ -9,6 +9,7 @@ Developers:
 from typing import Literal
 from typing import Optional
 
+from movai_core_shared.consts import LOGS_INFLUX_DB
 from movai_core_shared.consts import METRICS_INFLUX_DB
 from movai_core_shared.consts import PLATFORM_METRICS_INFLUX_DB
 from movai_core_shared.messages.general_data import Request
@@ -50,7 +51,28 @@ class QueryData(BaseModel):
 class MetricQueryData(BaseModel):
     measurement: str
     query_data: QueryData
+    count_field: Optional[str] = None
+    db_name: Literal[
+        Literal[LOGS_INFLUX_DB], Literal[METRICS_INFLUX_DB], Literal[PLATFORM_METRICS_INFLUX_DB]
+    ] = None
 
 
 class MetricQueryRequest(Request):
     req_data: MetricQueryData
+
+
+class MetricQueryResponse(BaseModel):
+    """Respone for MetricQueryRequest.
+
+    Attributes:
+        success (bool): Indicates if the query was successful.
+        results (dict): The results of the query.
+        error (Optional[str]): Error message if the query failed.
+        reason (Optional[str]): Reason for failure, if applicable.
+
+    """
+
+    success: bool
+    results: dict
+    error: Optional[str] = None
+    reason: Optional[str] = None