+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# vTBD
+- [BP-1354](https://movai.atlassian.net/browse/BP-1354): Add db_name to MetricQueryData
+
 # v3.2.0
 - [BP-1354](https://movai.atlassian.net/browse/BP-1354): Transitions metrics to influxdb